// QuickApp.Pro/ClientApp/app/models/asset-acquistion-type-audit.model.ts
export class AssetAcquistionTypeAudit {
    AssetAcquistionTypeAuditId: number;
    id: number;
    identification: string;
    name: string;
    memo: string;
    createdBy: string;
    updatedBy: string;
    createdDate: Date;
    updatedDate: Date;
    isDeleted: boolean;
    isActive: boolean;
}